% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnsens_example.R
\docType{data}
\name{bnsens_example}
\alias{bnsens_example}
\title{Example of Bayesian network}
\format{The Bayesian network \code{bnsens_example} comprehends the following nodes:
\itemize{
  \item \bold{y1}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y2}: three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
  \item \bold{y3}:  three-level factor with levels \emph{1}, \emph{2}, \emph{3}.
}}
\source{
Manuele Leonelli, Eva Riccomagno (2018).A geometric characterisation of sensitivity analysis in monomial models. \url{https://arxiv.org/abs/1901.02058}
}
\usage{
data(bnsens_example)
}
\description{
\code{bnsens_example} is a bn.fit object for a simple Bayesian network involving three variables.
}
\keyword{BNs}
