% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderp_covariation.R
\name{orderp_covar}
\alias{orderp_covar}
\title{Order-preserving co-variation scheme}
\usage{
orderp_covar(bnfit, node, value_node, value_parents, new_value)
}
\arguments{
\item{bnfit}{object of class bn.fit}

\item{node}{character string. Node of which the conditional probability distribution is being changed.}

\item{value_node}{character string. Level of \code{node}.}

\item{value_parents}{character string. Levels of \code{node}'s parents. The levels should be defined according to the order of the parents in \code{bnfit[[node]][["parents"]]}. If \code{node} has no parents, then should be set to \code{NULL}.}

\item{new_value}{numeric value between 0 and 1. Value to which the parameter should be updated.}
}
\description{
\code{orderp_covar} returns an updated Bayesian network using order-preserving
co-variation scheme
}
\details{
The Bayesian network on which parameter variation is being conducted should be expressed as a bn.fit object.
The name of the node to be varied, its level and its parent's level should be specified.
The parameter variation specified by the function is:\cr
\cr
 P ( \code{node} = \code{value_node} | parents = \code{value_parents} ) = \code{new_value} \cr
\cr
It should be noted that if two or more parameters in a distribution have the same value, the order is given by the one in the respective conditional probability table.
}
\seealso{
Other co-variation schemes: \code{\link{proportional_covar}},
  \code{\link{uniform_covar}}
}
\concept{co-variation schemes}
