% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_covar.R
\name{proportional_covar}
\alias{proportional_covar}
\title{Proportional co-variation scheme}
\usage{
proportional_covar(bnfit, node, value_node, value_parents, new_value)
}
\arguments{
\item{bnfit}{object of class bn.fit}

\item{node}{character string. Node of which the conditional probability distribution is being changed.}

\item{value_node}{character string. Level of \code{node}.}

\item{value_parents}{character string. Levels of \code{node}'s parents. The levels should be defined for the parents in the order found in \code{bnfit[[node]][["parents"]]}. If \code{node} has no parents, then should be set to \code{NULL}.}

\item{new_value}{numeric value between 0 and 1. Value to which the parameter should be updated.}
}
\description{
\code{proportional_covar} returns an updated Bayesian network using proportional co-variation scheme.
}
\details{
The Bayesian network on which parameter variation is being conducted should be expressed as a bn.fit object.
The name of the node to be varied, its level and its parent's level should be specified.
The parameter variation specified by the function is:\cr
\cr
 P ( \code{node} = \code{value_node} | parents = \code{value_parents} ) = \code{new_value} \cr
}
\seealso{
Other co-variation schemes: \code{\link{orderp_covar}},
  \code{\link{uniform_covar}}
}
\concept{co-variation schemes}
